requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
ajax==3.0.0
asgiref==3.7.2
asttokens==2.4.1
attrs==23.1.0
black==23.12.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
coverage==7.4.0
decorator==5.1.1
Django==5.0
django-appconf==1.0.6
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-yasg==1.21.7
executing==2.0.1
flake8==6.1.0
frozenlist==1.4.1
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
ipython==8.18.1
jedi==0.19.1
matplotlib-inline==0.1.6
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pyflakes==3.1.0
Pygments==2.17.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
stripe==7.9.0
traitlets==5.14.0
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
wcwidth==0.2.12
yarl==1.9.4
